markw: Add perf-profile0.conf
* Since lineageos 15.1 the power hal doesn't use the
perf_lock_use_profile() method anymore.
It was used to apply performance profiles.
On startup the perfd daemon is looking for at least
one profile which is used as default.
So there is only a perf-profiles0.conf to make the
daemon happy

* Removes log spam:
E ANDR-PERF-MPCTL: Invalid profile no. 0, total profiles 0 only